Improve some part of code
