+++ Sources/CoreBusiness/DataProvider/DataProvider.swift
@@ -36,11 +36,8 @@ final class DataProvider {
 
     func mediaMetadataPublisher(forUrn urn: String) -> AnyPublisher<MediaMetadata, Error> {
         mediaCompositionPublisher(forUrn: urn)
-            .tryMap { mediaCompositionResponse in
-                try MediaMetadata(
-                    mediaCompositionResponse: mediaCompositionResponse,
-                    dataProvider: self
-                )
+            .tryMap { response in
+                try MediaMetadata(mediaCompositionResponse: response, dataProvider: self)
             }
             .eraseToAnyPublisher()
     }
